coldata: change Batch interface to operate with uint64 length
This commit changes Batch interface to operate with uint64 length
because it unifies the way we store batches - now we will have a single
implementation MemBatch (with an exception for coldata.ZeroBatch). This
allows us to remove bufferedBatch.

Release note: None

+++ pkg/col/colserde/arrowbatchconverter.go
@@ -248,7 +248,7 @@ func (c *ArrowBatchConverter) ArrowToBatch(data []*array.Data, b coldata.Batch) 
 	// overwriting it. If the passed-in Batch is not suitable for use, a new one
 	// is allocated.
 	b.Reset(c.typs, n)
-	b.SetLength(uint16(n))
+	b.SetLength(uint64(n))
 	// Reset the batch, this resets the selection vector as well.
 	b.ResetInternalBatch()
